+++ src/core/engine/engine.py
@@ -4,8 +4,10 @@ import time
 from multiprocessing import Process, Queue, Value
 
 from src.core.config import Config
+from src.core.engine.enums import LOW_PRIORITY_ENVENT, MEDIUM_PRIORITY_ENVENT, HIGH_PRIORITY_ENVENT, LOW_PRIORITY_ENVENT_TIMEOUT, MEDIUM_PRIORITY_ENVENT_TIMEOUT, HIGH_PRIORITY_ENVENT_TIMEOUT
 from src.core.util.exceptions import EngineException
 from src.core.util.log import Logger
+from src.core.util.helper import utcnow_timestamp
 
 
 class EventEngine(object):
@@ -28,34 +30,60 @@ class EventEngine(object):
 
     # 执行事件循环
     def __run(self):
-        self.__logger.debug(
-            "src.core.engine.engine.EventEngine.__mainProcess.__run")
+        self.__logger.debug("src.core.engine.engine.EventEngine.__mainProcess.__run")
         while self.__active.value:
+            # 按优先级 获取队列中的事件 超时1秒
+            event = None
+            if not self.__highEventQueue.empty():
+                self.__logger.debug("src.core.engine.engine.EventEngine.__mainProcess.__run.__highEventQueue")
+                event = self.__highEventQueue.get(block=True, timeout=1))
+                while utcnow_timestamp()-int(event.timeStamp)>HIGH_PRIORITY_ENVENT_TIMEOUT:
+                    if not self.__highEventQueue.empty():
+                        self.__logger.warn("src.core.engine.engine.EventEngine.__mainProcess.__run.__highEventQueue TIMEOUT:"+event.type)
+                        event = self.__highEventQueue.get(block=True, timeout=1))
+                    else:
+                        event = None
+
+            if not self.__mediumEventQueue.empty() and event == None:
+                self.__logger.debug("src.core.engine.engine.EventEngine.__mainProcess.__run.__mediumEventQueue")
+                event = self.__mediumEventQueue.get(block=True, timeout=1))
+                while utcnow_timestamp()-int(event.timeStamp)>HIGH_PRIORITY_ENVENT_TIMEOUT:
+                    if not self.__mediumEventQueue.empty():
+                        self.__logger.warn("src.core.engine.engine.EventEngine.__mainProcess.__run.__mediumEventQueue TIMEOUT:"+event.type)
+                        event = self.__mediumEventQueue.get(block=True, timeout=1))
+                    else:
+                        event = None
+
+            if not self.__lowEnventQueue.empty() and event == None:
+                self.__logger.debug("src.core.engine.engine.EventEngine.__mainProcess.__run.__lowEnventQueue")
+                event = self.__lowEnventQueue.get(block=True, timeout=1))
+                while utcnow_timestamp()-int(event.timeStamp)>HIGH_PRIORITY_ENVENT_TIMEOUT:
+                    if not self.__lowEnventQueue.empty():
+                        self.__logger.warn("src.core.engine.engine.EventEngine.__mainProcess.__run.__lowEnventQueue TIMEOUT:"+event.type)
+                        event = self.__lowEnventQueue.get(block=True, timeout=1))
+                    else:
+                        event = None
             # 事件队列非空
-            if not self.__mediumEventQueue.empty():
-                # 获取队列中的事件 超时1秒
-                event = self.__mediumEventQueue.get(
-                    block=True, timeout=float(Config()._engine["timeout"]))
+            if not event == None:
                 # 执行事件
                 self.__logger.debug(
-                    "src.core.engine.engine.EventEngine.__mainProcess.__run.eventQueue: "
+                    "src.core.engine.engine.EventEngine.__mainProcess.__run.__eventQueue: "
                     + event.type)
                 self.__process(event)
             else:
                 # 等待 epoch
                 self.__logger.debug(
-                    "src.core.engine.engine.EventEngine.__mainProcess.__run.eventQueue: empty"
+                    "src.core.engine.engine.EventEngine.__mainProcess.__run.__eventQueue: empty"
                 )
                 time.sleep(float(Config()._engine["epoch"]))
         # 终止所有事件处理进程
-        self.__logger.debug(self.__processPool)
         for p in self.__processPool:
             p.join()
 
     # 执行事件
     def __process(self, event):
         self.__logger.debug(
-            "src.core.engine.engine.EventEngine.__mainProcess.__process: " +
+            "src.core.engine.engine.EventEngine.__mainProcess.__run.__process: " +
             event.type)
         if event.type in self.__handlers:
             for handler in self.__handlers[event.type]:
@@ -89,10 +117,9 @@ class EventEngine(object):
         self.__mainProcess.terminate()
 
     # 注册事件
-    def register(self, event, handler):
+    def register(self, type, handler):
         self.__logger.debug("src.core.engine.engine.EventEngine.register")
         # 尝试获取该事件类型对应的处理函数列表，若无则创建
-        type = event.type
         try:
             handlerList = self.__handlers[type]
         except KeyError:
@@ -102,10 +129,9 @@ class EventEngine(object):
             handlerList.append(handler)
         self.__handlers[type] = handlerList
 
-    def unregister(self, event, handler):
+    def unregister(self, type, handler):
         self.__logger.debug("src.core.engine.engine.EventEngine.unregister")
         # 尝试获取该事件类型对应的处理函数列表，若无则忽略该次注销请求
-        type = event.type
         try:
             handlerList = self.__handlers[type]
             # 如果该函数存在于列表中，则移除
@@ -122,7 +148,12 @@ class EventEngine(object):
     def sendEvent(self, event):
         self.__logger.debug("src.core.engine.engine.EventEngine.sendEvent")
         # 发送事件 像队列里存入事件
-        self.__mediumEventQueue.put(event)
+        if event.priority == LOW_PRIORITY_ENVENT:
+            self.__lowEnventQueue.put(event)
+        if event.priority == MEDIUM_PRIORITY_ENVENT:
+            self.__mediumEventQueue.put(event)
+        if event.priority == HIGH_PRIORITY_ENVENT:
+            self.__highEventQueue.put(event)
 
 
 class Event(object):
