+++ tests/engine/test_order.py
@@ -1,0 +1,18 @@
+# -*- coding: utf-8 -*-
+
+import os
+import sys
+sys.path.append(os.getcwd())
+
+from src.core.engine.engine import EventEngine
+from src.core.engine.listen import Listen, ListenHandler
+from src.core.util.util import Util
+from src.core.util.log import Logger
+
+# global var
+__eventEngine = EventEngine()
+__logger = Logger()
+
+# Begin Test
+if __name__ == '__main__':
+    pass
